+++ changelog.md
@@ -5,6 +5,7 @@ body {
 </style>
 
 ## TODO LIST (Clean at 2024.11.12)
+- [ ] gpr_test.cpp test_with_big_data function bug with ldlt
 - [x] Propagete the uncertainty.
 - [ ] Propagete the covariance using equation 35 in reference [1].
 - [ ] maybe realtime update model